Removes names code from Addresses tab
